+++ src/parser/ELR/builder/model/conflict.ts
@@ -12,7 +12,7 @@ export interface TempPartialConflict<T> {
   /** If this is a R-S conflict, this rule is a shifter rule. If this is a R-R conflict, this rule is a reducer rule. */
   anotherRule: TempGrammarRule<T>;
   /** A list of temp grammars that will cause conflicts when appear at the next of input. */
-  next: TempGrammar[];
+  next: TempGrammar[] | "*";
   /** Whether to handle conflict if reach the end of input using `reject`. */
   handleEnd: boolean;
 }
@@ -29,7 +29,7 @@ export interface Conflict<T> {
   /** If this is a R-S conflict, this rule is a shifter rule. If this is a R-R conflict, this rule is a reducer rule. */
   anotherRule: GrammarRule<T>;
   /** A list of grammars that will cause conflicts when appear at the next of input. */
-  next: Grammar[];
+  next: Grammar[] | "*";
   /** Whether to handle conflict if reach the end of input using `reject`. */
   handleEnd: boolean;
   /** R-S conflict only. How many grammars are overlapped in rule. */